requirements.txt
streamlit==1.40.0
scipy==1.14.1
numpy==1.26.4
pandas==2.2.3
nltk==3.9.1
zstandard==0.23.0
scikit-learn==1.3.0
psutil==6.1.0
faiss-cpu==1.9.0
sentence_transformers
tf-keras==2.18.0